[SWM-363] Refactor : split recommendations into 2 components
